refactor(server): extract pure service logic into utils
