+++ packages/twenty-server/src/engine/workspace-manager/workspace-migration/workspace-migration-builder/builders/permission-flag/workspace-migration-permission-flag-actions-builder.service.ts
@@ -7,15 +7,15 @@ import { type FlatEntityUpdateValidationArgs } from 'src/engine/workspace-manage
 import { type UniversalFlatEntityValidationArgs } from 'src/engine/workspace-manager/workspace-migration/workspace-migration-builder/types/universal-flat-entity-validation-args.type';
 import { type UniversalFlatEntityValidationReturnType } from 'src/engine/workspace-manager/workspace-migration/workspace-migration-builder/types/universal-flat-entity-validation-result.type';
 import { WorkspaceEntityMigrationBuilderService } from 'src/engine/workspace-manager/workspace-migration/workspace-migration-builder/services/workspace-entity-migration-builder.service';
-import { FlatPermissionFlagValidatorService } from 'src/engine/workspace-manager/workspace-migration/workspace-migration-builder/validators/services/flat-permission-flag-validator.service';
+import { validateFlatPermissionFlagCreation } from 'src/engine/workspace-manager/workspace-migration/workspace-migration-builder/validators/utils/validate-flat-permission-flag-creation.util';
+import { validateFlatPermissionFlagDeletion } from 'src/engine/workspace-manager/workspace-migration/workspace-migration-builder/validators/utils/validate-flat-permission-flag-deletion.util';
+import { validateFlatPermissionFlagUpdate } from 'src/engine/workspace-manager/workspace-migration/workspace-migration-builder/validators/utils/validate-flat-permission-flag-update.util';
 
 @Injectable()
 export class WorkspaceMigrationPermissionFlagActionsBuilderService extends WorkspaceEntityMigrationBuilderService<
   typeof ALL_METADATA_NAME.permissionFlag
 > {
-  constructor(
-    private readonly flatPermissionFlagValidatorService: FlatPermissionFlagValidatorService,
-  ) {
+  constructor() {
     super(ALL_METADATA_NAME.permissionFlag);
   }
 
@@ -27,10 +27,7 @@ export class WorkspaceMigrationPermissionFlagActionsBuilderService extends Works
     typeof ALL_METADATA_NAME.permissionFlag,
     'create'
   > {
-    const validationResult =
-      this.flatPermissionFlagValidatorService.validateFlatPermissionFlagCreation(
-        args,
-      );
+    const validationResult = validateFlatPermissionFlagCreation(args);
 
     if (validationResult.errors.length > 0) {
       return {
@@ -59,10 +56,7 @@ export class WorkspaceMigrationPermissionFlagActionsBuilderService extends Works
     typeof ALL_METADATA_NAME.permissionFlag,
     'delete'
   > {
-    const validationResult =
-      this.flatPermissionFlagValidatorService.validateFlatPermissionFlagDeletion(
-        args,
-      );
+    const validationResult = validateFlatPermissionFlagDeletion(args);
 
     if (validationResult.errors.length > 0) {
       return {
@@ -91,10 +85,7 @@ export class WorkspaceMigrationPermissionFlagActionsBuilderService extends Works
     typeof ALL_METADATA_NAME.permissionFlag,
     'update'
   > {
-    const validationResult =
-      this.flatPermissionFlagValidatorService.validateFlatPermissionFlagUpdate(
-        args,
-      );
+    const validationResult = validateFlatPermissionFlagUpdate(args);
 
     if (validationResult.errors.length > 0) {
       return {
