--- packages/twenty-server/src/engine/workspace-manager/workspace-migration/workspace-migration-builder/validators/services/flat-view-filter-validator.service.ts
@@ -1,458 +1,0 @@
-import { Injectable } from '@nestjs/common';
-
-import { msg, t } from '@lingui/core/macro';
-import { isNonEmptyString } from '@sniptt/guards';
-import { ALL_METADATA_NAME } from 'twenty-shared/metadata';
-import { FieldMetadataType, type ViewFilterOperand } from 'twenty-shared/types';
-import {
-  getFilterOperandsForFilterableFieldType,
-  getFilterTypeFromFieldType,
-  getFilterValueValidationIssue,
-  isDefined,
-} from 'twenty-shared/utils';
-
-import { findFlatEntityByUniversalIdentifier } from 'src/engine/metadata-modules/flat-entity/utils/find-flat-entity-by-universal-identifier.util';
-import { getEffectiveFilterFieldType } from 'src/engine/metadata-modules/flat-field-metadata/utils/get-effective-filter-field-type.util';
-import { type FlatFieldMetadata } from 'src/engine/metadata-modules/flat-field-metadata/types/flat-field-metadata.type';
-import { getInvalidSelectFilterOptionValues } from 'src/engine/metadata-modules/flat-field-metadata/utils/get-invalid-select-filter-option-values.util';
-import { ViewFilterExceptionCode } from 'src/engine/metadata-modules/view-filter/exceptions/view-filter.exception';
-import { type ViewFilterValue } from 'src/engine/metadata-modules/view-filter/types/view-filter-value.type';
-import { isFieldMetadataEntityOfType } from 'src/engine/utils/is-field-metadata-of-type.util';
-import { type FailedFlatEntityValidation } from 'src/engine/workspace-manager/workspace-migration/workspace-migration-builder/builders/types/failed-flat-entity-validation.type';
-import { getEmptyFlatEntityValidationError } from 'src/engine/workspace-manager/workspace-migration/workspace-migration-builder/builders/utils/get-flat-entity-validation-error.util';
-import { type FlatEntityUpdateValidationArgs } from 'src/engine/workspace-manager/workspace-migration/workspace-migration-builder/types/universal-flat-entity-update-validation-args.type';
-import { type UniversalFlatEntityValidationArgs } from 'src/engine/workspace-manager/workspace-migration/workspace-migration-builder/types/universal-flat-entity-validation-args.type';
-
-@Injectable()
-export class FlatViewFilterValidatorService {
-  constructor() {}
-
-  validateFlatViewFilterCreation({
-    flatEntityToValidate: flatViewFilterToValidate,
-    optimisticFlatEntityMapsAndRelatedFlatEntityMaps: {
-      flatViewFilterMaps: optimisticFlatViewFilterMaps,
-      flatViewMaps,
-      flatFieldMetadataMaps,
-      flatViewFilterGroupMaps,
-    },
-  }: UniversalFlatEntityValidationArgs<
-    typeof ALL_METADATA_NAME.viewFilter
-  >): FailedFlatEntityValidation<'viewFilter', 'create'> {
-    const validationResult = getEmptyFlatEntityValidationError({
-      flatEntityMinimalInformation: {
-        universalIdentifier: flatViewFilterToValidate.universalIdentifier,
-      },
-      metadataName: 'viewFilter',
-      type: 'create',
-    });
-
-    const existingViewFilter = findFlatEntityByUniversalIdentifier({
-      universalIdentifier: flatViewFilterToValidate.universalIdentifier,
-      flatEntityMaps: optimisticFlatViewFilterMaps,
-    });
-
-    if (isDefined(existingViewFilter)) {
-      validationResult.errors.push({
-        code: ViewFilterExceptionCode.INVALID_VIEW_FILTER_DATA,
-        message: t`View filter with this universal identifier already exists`,
-        userFriendlyMessage: msg`View filter already exists`,
-      });
-    }
-
-    const referencedView = findFlatEntityByUniversalIdentifier({
-      universalIdentifier: flatViewFilterToValidate.viewUniversalIdentifier,
-      flatEntityMaps: flatViewMaps,
-    });
-
-    if (!isDefined(referencedView)) {
-      validationResult.errors.push({
-        code: ViewFilterExceptionCode.INVALID_VIEW_FILTER_DATA,
-        message: t`View not found`,
-        userFriendlyMessage: msg`View not found`,
-      });
-    }
-
-    const referencedFieldMetadata = findFlatEntityByUniversalIdentifier({
-      universalIdentifier:
-        flatViewFilterToValidate.fieldMetadataUniversalIdentifier,
-      flatEntityMaps: flatFieldMetadataMaps,
-    });
-
-    if (!isDefined(referencedFieldMetadata)) {
-      validationResult.errors.push({
-        code: ViewFilterExceptionCode.INVALID_VIEW_FILTER_DATA,
-        message: t`Field metadata not found`,
-        userFriendlyMessage: msg`Field metadata not found`,
-      });
-    } else {
-      let relationTargetFieldType: FieldMetadataType | undefined;
-
-      if (
-        isDefined(
-          flatViewFilterToValidate.relationTargetFieldMetadataUniversalIdentifier,
-        )
-      ) {
-        const relationTargetFieldMetadata = findFlatEntityByUniversalIdentifier(
-          {
-            universalIdentifier:
-              flatViewFilterToValidate.relationTargetFieldMetadataUniversalIdentifier,
-            flatEntityMaps: flatFieldMetadataMaps,
-          },
-        );
-
-        relationTargetFieldType = relationTargetFieldMetadata?.type;
-      }
-
-      const incompatibleOperandError = this.getIncompatibleOperandError({
-        operand: flatViewFilterToValidate.operand,
-        fieldType: referencedFieldMetadata.type,
-        subFieldName: flatViewFilterToValidate.subFieldName,
-        relationTargetFieldType,
-      });
-
-      if (isDefined(incompatibleOperandError)) {
-        validationResult.errors.push(incompatibleOperandError);
-      }
-
-      const invalidValueError = this.getInvalidValueError({
-        operand: flatViewFilterToValidate.operand,
-        fieldType: referencedFieldMetadata.type,
-        subFieldName: flatViewFilterToValidate.subFieldName,
-        relationTargetFieldType,
-        value: flatViewFilterToValidate.value,
-      });
-
-      if (isDefined(invalidValueError)) {
-        validationResult.errors.push(invalidValueError);
-      }
-
-      if (
-        isFieldMetadataEntityOfType(
-          referencedFieldMetadata,
-          FieldMetadataType.SELECT,
-        ) ||
-        isFieldMetadataEntityOfType(
-          referencedFieldMetadata,
-          FieldMetadataType.MULTI_SELECT,
-        )
-      ) {
-        const invalidSelectOptionError = this.getInvalidSelectOptionError({
-          referencedFieldMetadata,
-          operand: flatViewFilterToValidate.operand,
-          value: flatViewFilterToValidate.value,
-        });
-
-        if (isDefined(invalidSelectOptionError)) {
-          validationResult.errors.push(invalidSelectOptionError);
-        }
-      }
-    }
-
-    if (
-      isDefined(flatViewFilterToValidate.viewFilterGroupUniversalIdentifier)
-    ) {
-      const referencedViewFilterGroup = findFlatEntityByUniversalIdentifier({
-        universalIdentifier:
-          flatViewFilterToValidate.viewFilterGroupUniversalIdentifier,
-        flatEntityMaps: flatViewFilterGroupMaps,
-      });
-
-      if (!isDefined(referencedViewFilterGroup)) {
-        validationResult.errors.push({
-          code: ViewFilterExceptionCode.INVALID_VIEW_FILTER_DATA,
-          message: t`View filter group not found`,
-          userFriendlyMessage: msg`View filter group not found`,
-        });
-      }
-    }
-
-    return validationResult;
-  }
-
-  validateFlatViewFilterDeletion({
-    flatEntityToValidate: flatViewFilterToValidate,
-    optimisticFlatEntityMapsAndRelatedFlatEntityMaps: {
-      flatViewFilterMaps: optimisticFlatViewFilterMaps,
-    },
-  }: UniversalFlatEntityValidationArgs<
-    typeof ALL_METADATA_NAME.viewFilter
-  >): FailedFlatEntityValidation<'viewFilter', 'delete'> {
-    const validationResult = getEmptyFlatEntityValidationError({
-      flatEntityMinimalInformation: {
-        universalIdentifier: flatViewFilterToValidate.universalIdentifier,
-      },
-      metadataName: 'viewFilter',
-      type: 'delete',
-    });
-
-    const existingViewFilter = findFlatEntityByUniversalIdentifier({
-      universalIdentifier: flatViewFilterToValidate.universalIdentifier,
-      flatEntityMaps: optimisticFlatViewFilterMaps,
-    });
-
-    if (!isDefined(existingViewFilter)) {
-      validationResult.errors.push({
-        code: ViewFilterExceptionCode.VIEW_FILTER_NOT_FOUND,
-        message: t`View filter not found`,
-        userFriendlyMessage: msg`View filter not found`,
-      });
-
-      return validationResult;
-    }
-
-    return validationResult;
-  }
-
-  validateFlatViewFilterUpdate({
-    universalIdentifier,
-    flatEntityUpdate,
-    optimisticFlatEntityMapsAndRelatedFlatEntityMaps: {
-      flatViewFilterMaps: optimisticFlatViewFilterMaps,
-      flatFieldMetadataMaps,
-      flatViewFilterGroupMaps,
-    },
-  }: FlatEntityUpdateValidationArgs<
-    typeof ALL_METADATA_NAME.viewFilter
-  >): FailedFlatEntityValidation<'viewFilter', 'update'> {
-    const existingViewFilter = findFlatEntityByUniversalIdentifier({
-      universalIdentifier,
-      flatEntityMaps: optimisticFlatViewFilterMaps,
-    });
-
-    const validationResult = getEmptyFlatEntityValidationError({
-      flatEntityMinimalInformation: {
-        universalIdentifier,
-      },
-      metadataName: 'viewFilter',
-      type: 'update',
-    });
-
-    if (!isDefined(existingViewFilter)) {
-      validationResult.errors.push({
-        code: ViewFilterExceptionCode.VIEW_FILTER_NOT_FOUND,
-        message: t`View filter not found`,
-        userFriendlyMessage: msg`View filter not found`,
-      });
-
-      return validationResult;
-    }
-
-    const updatedFlatViewFilter = {
-      ...existingViewFilter,
-      ...flatEntityUpdate,
-    };
-
-    const referencedFieldMetadata = findFlatEntityByUniversalIdentifier({
-      universalIdentifier:
-        updatedFlatViewFilter.fieldMetadataUniversalIdentifier,
-      flatEntityMaps: flatFieldMetadataMaps,
-    });
-
-    if (!isDefined(referencedFieldMetadata)) {
-      validationResult.errors.push({
-        code: ViewFilterExceptionCode.INVALID_VIEW_FILTER_DATA,
-        message: t`Field metadata not found`,
-        userFriendlyMessage: msg`Field metadata not found`,
-      });
-    } else {
-      let relationTargetFieldType: FieldMetadataType | undefined;
-
-      if (
-        isDefined(
-          updatedFlatViewFilter.relationTargetFieldMetadataUniversalIdentifier,
-        )
-      ) {
-        const relationTargetFieldMetadata = findFlatEntityByUniversalIdentifier(
-          {
-            universalIdentifier:
-              updatedFlatViewFilter.relationTargetFieldMetadataUniversalIdentifier,
-            flatEntityMaps: flatFieldMetadataMaps,
-          },
-        );
-
-        relationTargetFieldType = relationTargetFieldMetadata?.type;
-      }
-
-      const incompatibleOperandError = this.getIncompatibleOperandError({
-        operand: updatedFlatViewFilter.operand,
-        fieldType: referencedFieldMetadata.type,
-        subFieldName: updatedFlatViewFilter.subFieldName,
-        relationTargetFieldType,
-      });
-
-      if (isDefined(incompatibleOperandError)) {
-        validationResult.errors.push(incompatibleOperandError);
-      }
-
-      if (
-        'value' in flatEntityUpdate ||
-        'fieldMetadataUniversalIdentifier' in flatEntityUpdate ||
-        'operand' in flatEntityUpdate ||
-        'subFieldName' in flatEntityUpdate ||
-        'relationTargetFieldMetadataUniversalIdentifier' in flatEntityUpdate
-      ) {
-        const invalidValueError = this.getInvalidValueError({
-          operand: updatedFlatViewFilter.operand,
-          fieldType: referencedFieldMetadata.type,
-          subFieldName: updatedFlatViewFilter.subFieldName,
-          relationTargetFieldType,
-          value: updatedFlatViewFilter.value,
-        });
-
-        if (isDefined(invalidValueError)) {
-          validationResult.errors.push(invalidValueError);
-        }
-      }
-
-      if (
-        ('value' in flatEntityUpdate ||
-          'fieldMetadataUniversalIdentifier' in flatEntityUpdate ||
-          'operand' in flatEntityUpdate) &&
-        (isFieldMetadataEntityOfType(
-          referencedFieldMetadata,
-          FieldMetadataType.SELECT,
-        ) ||
-          isFieldMetadataEntityOfType(
-            referencedFieldMetadata,
-            FieldMetadataType.MULTI_SELECT,
-          ))
-      ) {
-        const invalidSelectOptionError = this.getInvalidSelectOptionError({
-          referencedFieldMetadata,
-          operand: updatedFlatViewFilter.operand,
-          value: updatedFlatViewFilter.value,
-        });
-
-        if (isDefined(invalidSelectOptionError)) {
-          validationResult.errors.push(invalidSelectOptionError);
-        }
-      }
-    }
-
-    if (isDefined(updatedFlatViewFilter.viewFilterGroupUniversalIdentifier)) {
-      const referencedViewFilterGroup = findFlatEntityByUniversalIdentifier({
-        universalIdentifier:
-          updatedFlatViewFilter.viewFilterGroupUniversalIdentifier,
-        flatEntityMaps: flatViewFilterGroupMaps,
-      });
-
-      if (!isDefined(referencedViewFilterGroup)) {
-        validationResult.errors.push({
-          code: ViewFilterExceptionCode.INVALID_VIEW_FILTER_DATA,
-          message: t`View filter group not found`,
-          userFriendlyMessage: msg`View filter group not found`,
-        });
-      }
-    }
-
-    return validationResult;
-  }
-
-  private getInvalidSelectOptionError({
-    referencedFieldMetadata,
-    operand,
-    value,
-  }: {
-    referencedFieldMetadata: Pick<
-      FlatFieldMetadata<
-        FieldMetadataType.SELECT | FieldMetadataType.MULTI_SELECT
-      >,
-      'type' | 'options' | 'label'
-    >;
-    operand: ViewFilterOperand;
-    value: ViewFilterValue;
-  }) {
-    const invalidValues = getInvalidSelectFilterOptionValues({
-      fieldMetadata: referencedFieldMetadata,
-      operand,
-      value,
-    });
-
-    if (invalidValues.length === 0) {
-      return undefined;
-    }
-
-    const invalidValuesText = invalidValues.join(', ');
-    const allowedValuesText = referencedFieldMetadata.options
-      ?.map((option) => option.value)
-      .join(', ');
-
-    return {
-      code: ViewFilterExceptionCode.INVALID_VIEW_FILTER_DATA,
-      message: t`Filter on "${referencedFieldMetadata.label}" uses option(s) ${invalidValuesText} that do not exist. Allowed values: ${allowedValuesText}.`,
-      userFriendlyMessage: msg`Filter uses a select option that does not exist`,
-    };
-  }
-
-  private getInvalidValueError({
-    operand,
-    fieldType,
-    subFieldName,
-    relationTargetFieldType,
-    value,
-  }: {
-    operand: ViewFilterOperand;
-    fieldType: FieldMetadataType;
-    subFieldName: string | null | undefined;
-    relationTargetFieldType: FieldMetadataType | undefined;
-    value: ViewFilterValue;
-  }) {
-    const issue = getFilterValueValidationIssue({
-      fieldType: getEffectiveFilterFieldType({
-        fieldType,
-        relationTargetFieldType,
-      }),
-      operand,
-      subFieldName,
-      value,
-    });
-
-    if (!isDefined(issue)) {
-      return undefined;
-    }
-
-    const { stringifiedValue, filterType, hint } = issue;
-
-    return {
-      code: ViewFilterExceptionCode.INVALID_VIEW_FILTER_DATA,
-      message: isNonEmptyString(hint)
-        ? t`Value "${stringifiedValue}" is not valid for operand "${operand}" on field type "${filterType}". ${hint}`
-        : t`Value "${stringifiedValue}" is not valid for operand "${operand}" on field type "${filterType}".`,
-      userFriendlyMessage: msg`Filter value is not valid for this operand`,
-    };
-  }
-
-  private getIncompatibleOperandError({
-    operand,
-    fieldType,
-    subFieldName,
-    relationTargetFieldType,
-  }: {
-    operand: ViewFilterOperand;
-    fieldType: FieldMetadataType;
-    subFieldName: string | null | undefined;
-    relationTargetFieldType: FieldMetadataType | undefined;
-  }) {
-    const effectiveFieldType = getEffectiveFilterFieldType({
-      fieldType,
-      relationTargetFieldType,
-    });
-
-    const filterType = getFilterTypeFromFieldType(effectiveFieldType);
-
-    const allowedOperands = getFilterOperandsForFilterableFieldType({
-      filterType,
-      subFieldName,
-    });
-
-    if (allowedOperands.includes(operand)) {
-      return undefined;
-    }
-
-    return {
-      code: ViewFilterExceptionCode.INVALID_VIEW_FILTER_DATA,
-      message: t`Operand "${operand}" is not supported on field type "${filterType}". Supported operands: ${allowedOperands.join(', ')}.`,
-      userFriendlyMessage: msg`Filter operand is not supported for this field type`,
-    };
-  }
-}
