--- packages/twenty-server/src/engine/workspace-manager/workspace-migration/workspace-migration-builder/validators/services/flat-view-validator.service.ts
@@ -1,574 +1,0 @@
-import { msg, t } from '@lingui/core/macro';
-import { type ALL_METADATA_NAME } from 'twenty-shared/metadata';
-import {
-  FeatureFlagKey,
-  FieldMetadataType,
-  RelationType,
-  ViewCalendarLayout,
-  ViewType,
-} from 'twenty-shared/types';
-import { getViewLayoutFromViewType, isDefined } from 'twenty-shared/utils';
-
-import { findFlatEntityByUniversalIdentifier } from 'src/engine/metadata-modules/flat-entity/utils/find-flat-entity-by-universal-identifier.util';
-import { isMorphOrRelationUniversalFlatFieldMetadata } from 'src/engine/metadata-modules/flat-field-metadata/utils/is-morph-or-relation-flat-field-metadata.util';
-import { ViewExceptionCode } from 'src/engine/metadata-modules/view/exceptions/view.exception';
-import { type AllUniversalFlatEntityMaps } from 'src/engine/workspace-manager/workspace-migration/universal-flat-entity/types/all-universal-flat-entity-maps.type';
-import { type UniversalFlatFieldMetadata } from 'src/engine/workspace-manager/workspace-migration/universal-flat-entity/types/universal-flat-field-metadata.type';
-import { type UniversalFlatView } from 'src/engine/workspace-manager/workspace-migration/universal-flat-entity/types/universal-flat-view.type';
-import {
-  type FailedFlatEntityValidation,
-  type FlatEntityValidationError,
-} from 'src/engine/workspace-manager/workspace-migration/workspace-migration-builder/builders/types/failed-flat-entity-validation.type';
-import { getEmptyFlatEntityValidationError } from 'src/engine/workspace-manager/workspace-migration/workspace-migration-builder/builders/utils/get-flat-entity-validation-error.util';
-import { type FlatEntityUpdateValidationArgs } from 'src/engine/workspace-manager/workspace-migration/workspace-migration-builder/types/universal-flat-entity-update-validation-args.type';
-import { type UniversalFlatEntityValidationArgs } from 'src/engine/workspace-manager/workspace-migration/workspace-migration-builder/types/universal-flat-entity-validation-args.type';
-
-export class FlatViewValidatorService {
-  constructor() {}
-
-  private validateCalendarFields({
-    flatView,
-    flatFieldMetadataMaps,
-    isCalendarWeekViewEnabled,
-  }: {
-    flatView: UniversalFlatView;
-    flatFieldMetadataMaps: AllUniversalFlatEntityMaps['flatFieldMetadataMaps'];
-    isCalendarWeekViewEnabled: boolean;
-  }): FlatEntityValidationError[] {
-    if (getViewLayoutFromViewType(flatView.type) !== ViewType.CALENDAR) {
-      return [];
-    }
-
-    const errors: FlatEntityValidationError[] = [];
-
-    if (!isDefined(flatView.calendarLayout)) {
-      errors.push({
-        code: ViewExceptionCode.INVALID_VIEW_DATA,
-        message: t`Calendar view must have a calendar layout`,
-        userFriendlyMessage: msg`Calendar view must have a calendar layout`,
-      });
-    }
-
-    // Widget calendars only allow non-month layouts when the day/week
-    // calendar feature is enabled; while it is off the month grid stays a
-    // data invariant rather than a UI-only convention.
-    if (
-      flatView.type === ViewType.CALENDAR_WIDGET &&
-      !isCalendarWeekViewEnabled &&
-      isDefined(flatView.calendarLayout) &&
-      flatView.calendarLayout !== ViewCalendarLayout.MONTH
-    ) {
-      errors.push({
-        code: ViewExceptionCode.INVALID_VIEW_DATA,
-        message: t`Calendar widget views only support the month layout`,
-        userFriendlyMessage: msg`Calendar widget views only support the month layout`,
-      });
-    }
-
-    if (!isDefined(flatView.calendarFieldMetadataUniversalIdentifier)) {
-      errors.push({
-        code: ViewExceptionCode.INVALID_VIEW_DATA,
-        message: t`Calendar view must have a calendar field`,
-        userFriendlyMessage: msg`Calendar view must have a calendar field`,
-      });
-
-      return errors;
-    }
-
-    const calendarFieldMetadata = findFlatEntityByUniversalIdentifier({
-      universalIdentifier: flatView.calendarFieldMetadataUniversalIdentifier,
-      flatEntityMaps: flatFieldMetadataMaps,
-    });
-
-    if (!isDefined(calendarFieldMetadata)) {
-      errors.push({
-        code: ViewExceptionCode.INVALID_VIEW_DATA,
-        message: t`Calendar field metadata not found`,
-        userFriendlyMessage: msg`Calendar field not found`,
-      });
-
-      return errors;
-    }
-
-    if (
-      calendarFieldMetadata.objectMetadataUniversalIdentifier !==
-      flatView.objectMetadataUniversalIdentifier
-    ) {
-      errors.push({
-        code: ViewExceptionCode.INVALID_VIEW_DATA,
-        message: t`Calendar field must belong to the view object`,
-        userFriendlyMessage: msg`Calendar field must belong to the view object`,
-      });
-    }
-
-    const calendarFieldIsDateKind =
-      calendarFieldMetadata.type === FieldMetadataType.DATE ||
-      calendarFieldMetadata.type === FieldMetadataType.DATE_TIME;
-
-    if (!calendarFieldIsDateKind) {
-      errors.push({
-        code: ViewExceptionCode.INVALID_VIEW_DATA,
-        message: t`Calendar field must be a date or date time field`,
-        userFriendlyMessage: msg`Calendar field must be a date or date time field`,
-      });
-    }
-
-    if (!isDefined(flatView.calendarEndFieldMetadataUniversalIdentifier)) {
-      return errors;
-    }
-
-    if (
-      flatView.calendarEndFieldMetadataUniversalIdentifier ===
-      flatView.calendarFieldMetadataUniversalIdentifier
-    ) {
-      errors.push({
-        code: ViewExceptionCode.INVALID_VIEW_DATA,
-        message: t`Calendar start and end fields must be different`,
-        userFriendlyMessage: msg`Calendar start and end fields must be different`,
-      });
-
-      return errors;
-    }
-
-    const calendarEndFieldMetadata = findFlatEntityByUniversalIdentifier({
-      universalIdentifier: flatView.calendarEndFieldMetadataUniversalIdentifier,
-      flatEntityMaps: flatFieldMetadataMaps,
-    });
-
-    if (!isDefined(calendarEndFieldMetadata)) {
-      errors.push({
-        code: ViewExceptionCode.INVALID_VIEW_DATA,
-        message: t`Calendar end field metadata not found`,
-        userFriendlyMessage: msg`Calendar end field not found`,
-      });
-
-      return errors;
-    }
-
-    if (
-      calendarEndFieldMetadata.objectMetadataUniversalIdentifier !==
-      flatView.objectMetadataUniversalIdentifier
-    ) {
-      errors.push({
-        code: ViewExceptionCode.INVALID_VIEW_DATA,
-        message: t`Calendar end field must belong to the view object`,
-        userFriendlyMessage: msg`Calendar end field must belong to the view object`,
-      });
-    }
-
-    const calendarEndFieldIsDateKind =
-      calendarEndFieldMetadata.type === FieldMetadataType.DATE ||
-      calendarEndFieldMetadata.type === FieldMetadataType.DATE_TIME;
-
-    if (!calendarEndFieldIsDateKind) {
-      errors.push({
-        code: ViewExceptionCode.INVALID_VIEW_DATA,
-        message: t`Calendar end field must be a date or date time field`,
-        userFriendlyMessage: msg`Calendar end field must be a date or date time field`,
-      });
-    } else if (
-      calendarFieldIsDateKind &&
-      calendarEndFieldMetadata.type !== calendarFieldMetadata.type
-    ) {
-      errors.push({
-        code: ViewExceptionCode.INVALID_VIEW_DATA,
-        message: t`Calendar start and end fields must have the same type`,
-        userFriendlyMessage: msg`Calendar start and end fields must have the same type`,
-      });
-    }
-
-    return errors;
-  }
-
-  private isAllowedKanbanMainGroupByField({
-    mainGroupByFieldMetadata,
-  }: {
-    mainGroupByFieldMetadata: UniversalFlatFieldMetadata;
-  }): boolean {
-    if (mainGroupByFieldMetadata.type === FieldMetadataType.SELECT) {
-      return true;
-    }
-
-    if (
-      !isMorphOrRelationUniversalFlatFieldMetadata(mainGroupByFieldMetadata)
-    ) {
-      return false;
-    }
-
-    return (
-      mainGroupByFieldMetadata.type === FieldMetadataType.RELATION &&
-      mainGroupByFieldMetadata.universalSettings?.relationType ===
-        RelationType.MANY_TO_ONE
-    );
-  }
-
-  public validateFlatViewUpdate({
-    universalIdentifier,
-    flatEntityUpdate,
-    optimisticFlatEntityMapsAndRelatedFlatEntityMaps: {
-      flatViewMaps: optimisticFlatViewMaps,
-      flatFieldMetadataMaps,
-    },
-    additionalCacheDataMaps: { featureFlagsMap },
-  }: FlatEntityUpdateValidationArgs<
-    typeof ALL_METADATA_NAME.view
-  >): FailedFlatEntityValidation<'view', 'update'> {
-    const existingFlatView = findFlatEntityByUniversalIdentifier({
-      universalIdentifier,
-      flatEntityMaps: optimisticFlatViewMaps,
-    });
-
-    const validationResult = getEmptyFlatEntityValidationError({
-      flatEntityMinimalInformation: {
-        universalIdentifier,
-      },
-      metadataName: 'view',
-      type: 'update',
-    });
-
-    if (!isDefined(existingFlatView)) {
-      validationResult.errors.push({
-        code: ViewExceptionCode.INVALID_VIEW_DATA,
-        message: t`View not found`,
-        userFriendlyMessage: msg`View not found`,
-      });
-
-      return validationResult;
-    }
-
-    const updatedFlatView: UniversalFlatView = {
-      ...existingFlatView,
-      ...flatEntityUpdate,
-    };
-
-    const kanbanAggregateOperationFieldMetadataUniversalIdentifierUpdate =
-      flatEntityUpdate.kanbanAggregateOperationFieldMetadataUniversalIdentifier;
-
-    if (
-      isDefined(
-        kanbanAggregateOperationFieldMetadataUniversalIdentifierUpdate,
-      ) &&
-      kanbanAggregateOperationFieldMetadataUniversalIdentifierUpdate !== null &&
-      !isDefined(
-        findFlatEntityByUniversalIdentifier({
-          universalIdentifier:
-            kanbanAggregateOperationFieldMetadataUniversalIdentifierUpdate,
-          flatEntityMaps: flatFieldMetadataMaps,
-        }),
-      )
-    ) {
-      validationResult.errors.push({
-        code: ViewExceptionCode.INVALID_VIEW_DATA,
-        message: t`View kanban aggregate field metadata not found`,
-        userFriendlyMessage: msg`View kanban aggregate field metadata not found`,
-      });
-    }
-
-    const viewBecomesKanban =
-      getViewLayoutFromViewType(updatedFlatView.type) === ViewType.KANBAN &&
-      getViewLayoutFromViewType(existingFlatView.type) !== ViewType.KANBAN;
-
-    if (viewBecomesKanban) {
-      if (
-        !isDefined(updatedFlatView.mainGroupByFieldMetadataUniversalIdentifier)
-      ) {
-        validationResult.errors.push({
-          code: ViewExceptionCode.INVALID_VIEW_DATA,
-          message: t`Kanban view must have a main group by field`,
-          userFriendlyMessage: msg`Kanban view must have a main group by field`,
-        });
-
-        return validationResult;
-      }
-
-      const mainGroupByFieldMetadata = findFlatEntityByUniversalIdentifier({
-        universalIdentifier:
-          updatedFlatView.mainGroupByFieldMetadataUniversalIdentifier,
-        flatEntityMaps: flatFieldMetadataMaps,
-      });
-
-      if (!isDefined(mainGroupByFieldMetadata)) {
-        validationResult.errors.push({
-          code: ViewExceptionCode.INVALID_VIEW_DATA,
-          message: t`Kanban main group by field metadata not found`,
-          userFriendlyMessage: msg`Kanban main group by field metadata not found`,
-        });
-      } else if (
-        !this.isAllowedKanbanMainGroupByField({
-          mainGroupByFieldMetadata,
-        })
-      ) {
-        validationResult.errors.push({
-          code: ViewExceptionCode.INVALID_VIEW_DATA,
-          message: t`Kanban main group by field must be a SELECT or a many-to-one relation field`,
-          userFriendlyMessage: msg`Kanban main group by field must be a select or a many-to-one relation field`,
-        });
-      }
-    }
-
-    const updatedMainGroupByFieldMetadataUniversalIdentifier =
-      updatedFlatView.mainGroupByFieldMetadataUniversalIdentifier;
-
-    const mainGroupByFieldMetadataIsAddedOrUpdated =
-      isDefined(updatedMainGroupByFieldMetadataUniversalIdentifier) &&
-      existingFlatView.mainGroupByFieldMetadataUniversalIdentifier !==
-        updatedMainGroupByFieldMetadataUniversalIdentifier;
-
-    if (mainGroupByFieldMetadataIsAddedOrUpdated && !viewBecomesKanban) {
-      const mainGroupByFieldMetadata = findFlatEntityByUniversalIdentifier({
-        universalIdentifier: updatedMainGroupByFieldMetadataUniversalIdentifier,
-        flatEntityMaps: flatFieldMetadataMaps,
-      });
-
-      if (!isDefined(mainGroupByFieldMetadata)) {
-        validationResult.errors.push({
-          code: ViewExceptionCode.INVALID_VIEW_DATA,
-          message: t`Kanban main group by field metadata not found`,
-          userFriendlyMessage: msg`Kanban main group by field metadata not found`,
-        });
-      } else if (
-        !this.isAllowedKanbanMainGroupByField({
-          mainGroupByFieldMetadata,
-        })
-      ) {
-        validationResult.errors.push({
-          code: ViewExceptionCode.INVALID_VIEW_DATA,
-          message: t`Kanban main group by field must be a SELECT or a many-to-one relation field`,
-          userFriendlyMessage: msg`Kanban main group by field must be a select or a many-to-one relation field`,
-        });
-      }
-    }
-
-    validationResult.errors.push(
-      ...this.validateCalendarFields({
-        flatView: updatedFlatView,
-        flatFieldMetadataMaps,
-        isCalendarWeekViewEnabled:
-          featureFlagsMap[FeatureFlagKey.IS_CALENDAR_WEEK_VIEW_ENABLED],
-      }),
-    );
-
-    return validationResult;
-  }
-
-  public validateFlatViewDeletion({
-    flatEntityToValidate,
-    optimisticFlatEntityMapsAndRelatedFlatEntityMaps: {
-      flatViewMaps: optimisticFlatViewMaps,
-      flatObjectMetadataMaps: optimisticFlatObjectMetadataMaps,
-    },
-  }: UniversalFlatEntityValidationArgs<
-    typeof ALL_METADATA_NAME.view
-  >): FailedFlatEntityValidation<'view', 'delete'> {
-    const validationResult = getEmptyFlatEntityValidationError({
-      flatEntityMinimalInformation: {
-        universalIdentifier: flatEntityToValidate.universalIdentifier,
-      },
-      metadataName: 'view',
-      type: 'delete',
-    });
-
-    const existingFlatView = findFlatEntityByUniversalIdentifier({
-      universalIdentifier: flatEntityToValidate.universalIdentifier,
-      flatEntityMaps: optimisticFlatViewMaps,
-    });
-
-    if (!isDefined(existingFlatView)) {
-      validationResult.errors.push({
-        code: ViewExceptionCode.INVALID_VIEW_DATA,
-        message: t`View not found`,
-        userFriendlyMessage: msg`View not found`,
-      });
-
-      return validationResult;
-    }
-
-    const parentObjectStillExists = isDefined(
-      findFlatEntityByUniversalIdentifier({
-        universalIdentifier: existingFlatView.objectMetadataUniversalIdentifier,
-        flatEntityMaps: optimisticFlatObjectMetadataMaps,
-      }),
-    );
-
-    if (parentObjectStillExists) {
-      const viewsForSameObject = Object.values(
-        optimisticFlatViewMaps.byUniversalIdentifier,
-      ).filter(
-        (view) =>
-          isDefined(view) &&
-          view.objectMetadataUniversalIdentifier ===
-            existingFlatView.objectMetadataUniversalIdentifier,
-      );
-
-      if (viewsForSameObject.length <= 1) {
-        validationResult.errors.push({
-          code: ViewExceptionCode.INVALID_VIEW_DATA,
-          message: t`Cannot delete the only view for this object`,
-          userFriendlyMessage: msg`Cannot delete the only view for this object`,
-        });
-      }
-    }
-
-    return validationResult;
-  }
-
-  public validateFlatViewCreation({
-    flatEntityToValidate: flatViewToValidate,
-    optimisticFlatEntityMapsAndRelatedFlatEntityMaps: {
-      flatViewMaps: optimisticFlatViewMaps,
-      flatFieldMetadataMaps,
-      flatObjectMetadataMaps,
-    },
-    additionalCacheDataMaps: { featureFlagsMap },
-  }: UniversalFlatEntityValidationArgs<
-    typeof ALL_METADATA_NAME.view
-  >): FailedFlatEntityValidation<'view', 'create'> {
-    const validationResult = getEmptyFlatEntityValidationError({
-      flatEntityMinimalInformation: {
-        universalIdentifier: flatViewToValidate.universalIdentifier,
-      },
-      metadataName: 'view',
-      type: 'create',
-    });
-
-    const optimisticFlatObjectMetadata = findFlatEntityByUniversalIdentifier({
-      universalIdentifier: flatViewToValidate.objectMetadataUniversalIdentifier,
-      flatEntityMaps: flatObjectMetadataMaps,
-    });
-
-    if (!isDefined(optimisticFlatObjectMetadata)) {
-      validationResult.errors.push({
-        code: ViewExceptionCode.INVALID_VIEW_DATA,
-        message: t`Object metadata not found`,
-        userFriendlyMessage: msg`Object metadata not found`,
-      });
-    }
-
-    if (
-      isDefined(
-        findFlatEntityByUniversalIdentifier({
-          universalIdentifier: flatViewToValidate.universalIdentifier,
-          flatEntityMaps: optimisticFlatViewMaps,
-        }),
-      )
-    ) {
-      validationResult.errors.push({
-        code: ViewExceptionCode.INVALID_VIEW_DATA,
-        message: t`View with same universal identifier already exists`,
-        userFriendlyMessage: msg`View already exists`,
-      });
-    }
-
-    // Every ViewKey value is reserved for an engine-owned singleton view per
-    // object (INDEX table view, FIELDS_WIDGET record-page view).
-    const reservedViewKey = flatViewToValidate.key;
-
-    if (isDefined(reservedViewKey)) {
-      if (flatViewToValidate.isSystemSideEffect !== true) {
-        validationResult.errors.push({
-          code: ViewExceptionCode.INVALID_VIEW_DATA,
-          message: t`The ${reservedViewKey} view key is reserved for the engine-owned default view; remove the key from the view definition`,
-          userFriendlyMessage: msg`The ${reservedViewKey} view key is reserved for the default view`,
-        });
-      }
-
-      const objectAlreadyHasFlatViewWithSameKey =
-        isDefined(optimisticFlatObjectMetadata) &&
-        optimisticFlatObjectMetadata.viewUniversalIdentifiers.some(
-          (viewUniversalIdentifier) => {
-            const flatView = findFlatEntityByUniversalIdentifier({
-              universalIdentifier: viewUniversalIdentifier,
-              flatEntityMaps: optimisticFlatViewMaps,
-            });
-
-            return (
-              isDefined(flatView) &&
-              flatView.key === reservedViewKey &&
-              !isDefined(flatView.deletedAt)
-            );
-          },
-        );
-
-      if (objectAlreadyHasFlatViewWithSameKey) {
-        validationResult.errors.push({
-          code: ViewExceptionCode.INVALID_VIEW_DATA,
-          message: t`Object already has a view with the ${reservedViewKey} key`,
-          userFriendlyMessage: msg`This object already has a default view`,
-        });
-      }
-    }
-
-    if (
-      isDefined(
-        flatViewToValidate.kanbanAggregateOperationFieldMetadataUniversalIdentifier,
-      ) &&
-      !isDefined(
-        findFlatEntityByUniversalIdentifier({
-          universalIdentifier:
-            flatViewToValidate.kanbanAggregateOperationFieldMetadataUniversalIdentifier,
-          flatEntityMaps: flatFieldMetadataMaps,
-        }),
-      )
-    ) {
-      validationResult.errors.push({
-        code: ViewExceptionCode.INVALID_VIEW_DATA,
-        message: t`View kanban aggregate field metadata not found`,
-        userFriendlyMessage: msg`View kanban aggregate field metadata not found`,
-      });
-    }
-
-    const isKanban =
-      getViewLayoutFromViewType(flatViewToValidate.type) === ViewType.KANBAN;
-
-    if (isKanban) {
-      if (
-        !isDefined(
-          flatViewToValidate.mainGroupByFieldMetadataUniversalIdentifier,
-        )
-      ) {
-        validationResult.errors.push({
-          code: ViewExceptionCode.INVALID_VIEW_DATA,
-          message: t`Kanban view must have a main group by field`,
-          userFriendlyMessage: msg`Kanban view must have a main group by field`,
-        });
-
-        return validationResult;
-      }
-
-      const mainGroupByFieldMetadata = findFlatEntityByUniversalIdentifier({
-        universalIdentifier:
-          flatViewToValidate.mainGroupByFieldMetadataUniversalIdentifier,
-        flatEntityMaps: flatFieldMetadataMaps,
-      });
-
-      if (!isDefined(mainGroupByFieldMetadata)) {
-        validationResult.errors.push({
-          code: ViewExceptionCode.INVALID_VIEW_DATA,
-          message: t`Kanban main group by field metadata not found`,
-          userFriendlyMessage: msg`Kanban main group by field metadata not found`,
-        });
-      } else if (
-        !this.isAllowedKanbanMainGroupByField({
-          mainGroupByFieldMetadata,
-        })
-      ) {
-        validationResult.errors.push({
-          code: ViewExceptionCode.INVALID_VIEW_DATA,
-          message: t`Kanban main group by field must be a SELECT or a many-to-one relation field`,
-          userFriendlyMessage: msg`Kanban main group by field must be a select or a many-to-one relation field`,
-        });
-      }
-    }
-
-    validationResult.errors.push(
-      ...this.validateCalendarFields({
-        flatView: flatViewToValidate,
-        flatFieldMetadataMaps,
-        isCalendarWeekViewEnabled:
-          featureFlagsMap[FeatureFlagKey.IS_CALENDAR_WEEK_VIEW_ENABLED],
-      }),
-    );
-
-    return validationResult;
-  }
-}
