+++ packages/twenty-server/src/engine/workspace-manager/workspace-migration/workspace-migration-builder/validators/utils/__tests__/is-allowed-kanban-main-group-by-field.util.spec.ts
@@ -1,0 +1,71 @@
+import { FieldMetadataType, RelationType } from 'twenty-shared/types';
+
+import { type UniversalFlatFieldMetadata } from 'src/engine/workspace-manager/workspace-migration/universal-flat-entity/types/universal-flat-field-metadata.type';
+import { isAllowedKanbanMainGroupByField } from 'src/engine/workspace-manager/workspace-migration/workspace-migration-builder/validators/utils/is-allowed-kanban-main-group-by-field.util';
+
+const buildField = (
+  field: Partial<UniversalFlatFieldMetadata>,
+): UniversalFlatFieldMetadata => field as UniversalFlatFieldMetadata;
+
+describe('isAllowedKanbanMainGroupByField', () => {
+  it('should allow a SELECT field', () => {
+    expect(
+      isAllowedKanbanMainGroupByField({
+        mainGroupByFieldMetadata: buildField({
+          type: FieldMetadataType.SELECT,
+        }),
+      }),
+    ).toBe(true);
+  });
+
+  it('should allow a MANY_TO_ONE relation field', () => {
+    expect(
+      isAllowedKanbanMainGroupByField({
+        mainGroupByFieldMetadata: buildField({
+          type: FieldMetadataType.RELATION,
+          universalSettings: { relationType: RelationType.MANY_TO_ONE },
+        }),
+      }),
+    ).toBe(true);
+  });
+
+  it('should reject a ONE_TO_MANY relation field', () => {
+    expect(
+      isAllowedKanbanMainGroupByField({
+        mainGroupByFieldMetadata: buildField({
+          type: FieldMetadataType.RELATION,
+          universalSettings: { relationType: RelationType.ONE_TO_MANY },
+        }),
+      }),
+    ).toBe(false);
+  });
+
+  it('should reject a relation field with no relation type set', () => {
+    expect(
+      isAllowedKanbanMainGroupByField({
+        mainGroupByFieldMetadata: buildField({
+          type: FieldMetadataType.RELATION,
+          universalSettings: {},
+        }),
+      }),
+    ).toBe(false);
+  });
+
+  it('should reject a plain scalar field', () => {
+    expect(
+      isAllowedKanbanMainGroupByField({
+        mainGroupByFieldMetadata: buildField({ type: FieldMetadataType.TEXT }),
+      }),
+    ).toBe(false);
+  });
+
+  it('should reject a MULTI_SELECT field', () => {
+    expect(
+      isAllowedKanbanMainGroupByField({
+        mainGroupByFieldMetadata: buildField({
+          type: FieldMetadataType.MULTI_SELECT,
+        }),
+      }),
+    ).toBe(false);
+  });
+});
