+++ packages/twenty-server/src/modules/workflow/workflow-builder/workflow-validation/utils/has-output-schema.util.ts
@@ -1,0 +1,28 @@
+import { isObject } from '@sniptt/guards';
+
+export const hasOutputSchema = (
+  settings:
+    | { outputSchema?: unknown; expectedOutputSchema?: unknown }
+    | null
+    | undefined,
+): boolean => {
+  const expectedOutputSchema = settings?.expectedOutputSchema;
+
+  if (
+    isObject(expectedOutputSchema) &&
+    Object.keys(expectedOutputSchema).length > 0
+  ) {
+    return true;
+  }
+
+  const outputSchema = settings?.outputSchema;
+
+  return (
+    isObject(outputSchema) &&
+    Object.keys(outputSchema).length > 0 &&
+    !(
+      '_outputSchemaType' in outputSchema &&
+      outputSchema._outputSchemaType === 'LINK'
+    )
+  );
+};
